requirements.txt
groq>=0.12.0
google-play-scraper>=1.2.7
pandas>=2.1.0
python-dotenv>=1.0.0
streamlit>=1.38.0
markdown2>=2.5.0
typer>=0.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
apscheduler>=3.10.1
pytz>=2023.3
